onchain_parser/pyproject.toml
[project]
name = "onchain-parser"
version = "0.1.0"
description = "Python-first Bitcoin on-chain analytics parser"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.40",
  "alembic>=1.16.0",
  "psycopg[binary]>=3.2.0",
  "python-bitcoinrpc>=1.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.12.0",
  "mypy>=1.16.0",
]

[tool.ruff]
line-length = 100

target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["app"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
